[cluster-refactor]split DataGroupServiceImpls into engine and impls
